requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.33.2
sentence-transformers==2.2.2
huggingface-hub==0.18.0
faiss-cpu==1.7.4
streamlit==1.28.0
python-dotenv==1.0.0
langchain-community==0.0.38
tiktoken==0.4.0
requests==2.31.0
beautifulsoup4==4.14.3
lxml==4.9.3
html5lib==1.1
scikit-learn==1.3.2
numpy==1.26.4
scipy==1.11.1
google-search-results==2.4.2
